requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
tencentcloud-sdk-python==3.0.950
urllib3==2.0.4
